more singed/unsigned fixes
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34202 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/count.c
@@ -781,7 +781,7 @@ static void mark_region(Sentence sent,
 
 void delete_unmarked_disjuncts(Sentence sent)
 {
-	int w;
+	size_t w;
 	Disjunct *d_head, *d, *dx;
 
 	for (w=0; w<sent->length; w++) {
